pyproject.toml
[project]
name = "QaaS"
version = "0.1.0"
description = "Programming task for Oper"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "channels>=4.3.1",
    "channels-redis>=4.3.0",
    "daphne>=4.2.1",
    "django<5",
    "django-cors-headers>=4.7.0",
    "django-nested-admin>=4.1.1",
    "djangorestframework>=3.16.0",
    "psycopg[binary]>=3.2.6",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.1.0",
    "pytest-django>=4.11.1",
    "ruff>=0.11.7",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "oper.settings"
